+++ src/Ratchet/Wamp/WampConnection.php
@@ -2,7 +2,7 @@
 namespace Ratchet\Wamp;
 use Ratchet\ConnectionInterface;
 use Ratchet\AbstractConnectionDecorator;
-use Ratchet\Wamp\WampServer as WAMP;
+use Ratchet\Wamp\ServerProtocol as WAMP;
 
 /**
  * A ConnectionInterface object wrapper that is passed to your WAMP application
@@ -10,6 +10,9 @@ use Ratchet\Wamp\WampServer as WAMP;
  * @property stdClass $WAMP
  */
 class WampConnection extends AbstractConnectionDecorator {
+    /**
+     * {@inheritdoc}
+     */
     public function __construct(ConnectionInterface $conn) {
         parent::__construct($conn);
 
@@ -21,6 +24,7 @@ class WampConnection extends AbstractConnectionDecorator {
     }
 
     /**
+     * Successfully respond to a call made by the client
      * @param string The unique ID given by the client to respond to
      * @param array An array of data to return to the client
      */
@@ -29,13 +33,14 @@ class WampConnection extends AbstractConnectionDecorator {
     }
 
     /**
+     * Respond with an error to a client call
      * @param string The unique ID given by the client to respond to
      * @param string The URI given by the client ot respond to
      * @param string A developer-oriented description of the error
      * @param string|null An optional human readable detail message to send back
      */
     public function callError($id, $topic, $desc = '', $details = null) {
-        $data = array(WAMP::MSG_CALL_ERROR, $id, $topic, $desc);
+        $data = array(WAMP::MSG_CALL_ERROR, $id, (string)$topic, $desc);
 
         if (null !== $details) {
             $data[] = $details;
@@ -49,7 +54,7 @@ class WampConnection extends AbstractConnectionDecorator {
      * @param mixed Data to send with the event.  Anything that is json'able
      */
     public function event($topic, $msg) {
-        $this->send(json_encode(array(WAMP::MSG_EVENT, $topic, $msg)));
+        $this->send(json_encode(array(WAMP::MSG_EVENT, (string)$topic, $msg)));
     }
 
     /**
@@ -57,8 +62,8 @@ class WampConnection extends AbstractConnectionDecorator {
      * @param string
      */
     public function prefix($curie, $uri) {
-        $this->WAMP->prefixes[$curie] = $uri;
-        $this->send(json_encode(array(WAMP::MSG_PREFIX, $curie, $uri)));
+        $this->WAMP->prefixes[$curie] = (string)$uri;
+        $this->send(json_encode(array(WAMP::MSG_PREFIX, $curie, (string)$uri)));
     }
 
     /**